Parser and lexer
Lexer fixed a bug.

Parser together with executor. Will separate later. Just for ease of testing atm.

Could get pwd to work. Everything else not yet.

The idea is it reads the first exe_tok-> cmd  which would be a one of the commands (echo, pwd, etc). It will then increment and start reading the rest as the args as it increment.